Update script.js
Kept somewhat of the fishCheck function, but did some editing to it-- Making the fish go bad when the number three is computed; If not(if the number computed is lower than three), then result in the fish being good. The number could be raised higher, but I thought this was good enough.