Fight!
How about this? This is what I thought of; Nest a switch statement inside of the previous switch statement, then add more cases for the user (with a new variable included) - 'fight' or 'run'. If the user's choice is one of those, then it'll produce one of the following codes. I'm probably missing something, but I just wanted to give this a shot.